Move the “…” menu to the budget name
