+++ packages/loot-design/src/components/sidebar.js
@@ -447,62 +447,6 @@ function ToggleButton({ style, onFloat }) {
   );
 }
 
-const MenuButton = withRouter(function MenuButton({ history }) {
-  let dispatch = useDispatch();
-  let [menuOpen, setMenuOpen] = useState(false);
-
-  function onMenuSelect(type) {
-    setMenuOpen(false);
-
-    switch (type) {
-      case 'settings':
-        history.push('/settings');
-        break;
-      case 'help':
-        window.open('https://actualbudget.github.io/docs', '_blank');
-        break;
-      case 'close':
-        dispatch(closeBudget());
-        break;
-      default:
-    }
-  }
-
-  let items = [
-    { name: 'settings', text: 'Settings' },
-    { name: 'help', text: 'Help' },
-    { name: 'close', text: 'Close File' }
-  ];
-
-  return (
-    <Button
-      bare
-      style={{
-        color: colors.n5,
-        flexShrink: 0
-      }}
-      activeStyle={{ color: colors.p7 }}
-      onClick={() => setMenuOpen(true)}
-      aria-label="Menu"
-    >
-      <DotsHorizontalTriple
-        width={15}
-        height={15}
-        style={{ color: 'inherit', transform: 'rotateZ(0deg)' }}
-      />
-      {menuOpen && (
-        <Tooltip
-          position="bottom-right"
-          style={{ padding: 0 }}
-          onClose={() => setMenuOpen(false)}
-        >
-          <Menu onMenuSelect={onMenuSelect} items={items} />
-        </Tooltip>
-      )}
-    </Button>
-  );
-});
-
 function Tools() {
   let [isOpen, setOpen] = useState(false);
   let onToggle = useCallback(() => setOpen(open => !open), []);
@@ -645,7 +589,6 @@ export function Sidebar({
         <View style={{ flex: 1, flexDirection: 'row' }} />
 
         {!hasWindowButtons && <ToggleButton onFloat={onFloat} />}
-        {Platform.isBrowser && <MenuButton />}
       </View>
 
       <View style={{ overflow: 'auto' }}>
